+++ fs/vfs/fs_dup3.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * fs/vfs/fs_dup2.c
+ * fs/vfs/fs_dup3.c
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -34,12 +34,7 @@
 #include <errno.h>
 #include <fcntl.h>
 
-#include "sched/sched.h"
 #include "inode/inode.h"
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
 
 /****************************************************************************
  * Name: file_dup3
@@ -85,7 +80,7 @@ int file_dup3(FAR struct file *filep1, FAR struct file *filep2, int flags)
 
   /* If there is already an inode contained in the new file structure,
    * close the file and release the inode.
-   * But we need keep the filep2->f_inode, incase of realloced by others.
+   * But we need keep the filep2->f_inode, in case of realloced by others.
    */
 
   ret = file_close_without_clear(filep2);
@@ -178,64 +173,19 @@ int file_dup3(FAR struct file *filep1, FAR struct file *filep2, int flags)
 }
 
 /****************************************************************************
- * Name: file_dup2
- *
- * Description:
- *   Assign an inode to a specific files structure.  This is the heart of
- *   dup2.
- *
- *   Equivalent to the non-standard dup2() function except that it
- *   accepts struct file instances instead of file descriptors and it does
- *   not set the errno variable.
- *
- * Returned Value:
- *   Zero (OK) is returned on success; a negated errno value is return on
- *   any failure.
- *
- ****************************************************************************/
-
-int file_dup2(FAR struct file *filep1, FAR struct file *filep2)
-{
-  return file_dup3(filep1, filep2, 0);
-}
-
-/****************************************************************************
- * Name: nx_dup2
- *
- * Description:
- *   nx_dup2() is similar to the standard 'dup2' interface except that is
- *   not a cancellation point and it does not modify the errno variable.
- *
- *   nx_dup2() is an internal NuttX interface and should not be called from
- *   applications.
- *
- *   Clone a file descriptor to a specific descriptor number.
- *
- * Returned Value:
- *   fd2 is returned on success; a negated errno value is return on
- *   any failure.
- *
- ****************************************************************************/
-
-int nx_dup2(int fd1, int fd2)
-{
-  return nx_dup2_from_tcb(this_task(), fd1, fd2);
-}
-
-/****************************************************************************
- * Name: dup2
+ * Name: dup3
  *
  * Description:
  *   Clone a file descriptor or socket descriptor to a specific descriptor
- *   number
+ *   number and specific flags.
  *
  ****************************************************************************/
 
-int dup2(int fd1, int fd2)
+int dup3(int fd1, int fd2, int flags)
 {
   int ret;
 
-  ret = nx_dup2(fd1, fd2);
+  ret = nx_dup3_from_tcb(this_task(), fd1, fd2, flags);
   if (ret < 0)
     {
       set_errno(-ret);
